added get recipe by id functionality to view model

+++ app/src/main/java/uk/ac/bbk/dcs/mobile_computing_project_rroman08/RecipeDao.kt
@@ -14,7 +14,7 @@ interface RecipeDao {
 
     // Retrieve a recipe by ID from db
     @Query("SELECT * FROM recipes WHERE id = :id")
-    suspend fun getRecipeById(id: Int): Recipe?
+    suspend fun getRecipeById(id: Long): Recipe?
 
     // Insert a new recipe into db
     @Insert()
